Fixed when keyword and dropped register keyword from duties variable
The duties variable is a list of tasks whose items a Ansible loop
will feed to module jm1.ansible.execute_module. Because of this loop,
the register keyword in the duties variable will not work the same way
as Ansible's register keyword being applied to a single (non-loop)
task. Keyword register in duties is in fact useless.